refactor: Refactor post forms and tag selectors for improved functionality and maintainability
