+++ resources/js/pages/posts/post.tsx
@@ -55,7 +55,7 @@ export default function PostPage({ post }: PostProps) {
                     <div className="mt-2 flex flex-wrap gap-2">
                         {post.tags.map((tag, index) => {
                             return (
-                                <Badge key={index} variant="secondary" className="cursor-pointer">
+                                <Badge key={index} variant="secondary">
                                     {tag.name}
                                 </Badge>
                             );
